+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatInternalLog.cs
@@ -8,6 +8,7 @@ namespace Unity.Android.Logcat
 {
     internal class AndroidLogcatInternalLog : EditorWindow
     {
+        const int kMaxInternalLogBuffer = 15000;
         static AndroidLogcatInternalLog ms_Instance = null;
         static StringBuilder ms_LogEntries = new StringBuilder();
 
@@ -30,13 +31,27 @@ namespace Unity.Android.Logcat
         /// <param name="args"></param>
         public static void Log(string message, params object[] args)
         {
+            Log(string.Format(message, args));
+        }
+
+        public static void Log(string message)
+        {
             lock (ms_LogEntries)
             {
                 var timedMessage = AndroidLogcatDispatcher.isMainThread ? "[MainThread]" : "[WorkThread] ";
-                timedMessage += DateTime.Now.ToString("HH:mm:ss.ffff") + " " + string.Format(message, args);
+                timedMessage += DateTime.Now.ToString("HH:mm:ss.ffff") + " " + message;
                 ms_LogEntries.AppendLine(timedMessage);
 
+                const int MaxTriesToStripBuffer = 30;
+                for (int i = 0; i < MaxTriesToStripBuffer; i++)
+                {
+                    if (ms_LogEntries.Length < kMaxInternalLogBuffer)
+                        break;
+                    ms_LogEntries.Remove(0, Convert.ToString(ms_LogEntries).Split('\n').FirstOrDefault().Length + 1);
+                }
+
                 Console.WriteLine("[Logcat] " + timedMessage);
+
             }
 
             if (AndroidLogcatDispatcher.isMainThread && ms_Instance != null)