+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatRuntime.cs
@@ -53,7 +53,8 @@ namespace Unity.Android.Logcat
             get { ValidateIsInitialized(); return m_ScreenRecorder; }
         }
 
-        public abstract AndroidLogcatMessageProviderBase CreateMessageProvider(AndroidBridge.ADB adb, string filter, Priority priority, int packageID, string logPrintFormat, IAndroidLogcatDevice device, Action<string> logCallbackAction);
+
+        public abstract AndroidLogcatMessageProviderBase CreateMessageProvider(AndroidBridge.ADB adb, Priority priority, int packageID, string logPrintFormat, IAndroidLogcatDevice device, Action<string> logCallbackAction);
         protected abstract AndroidLogcatDeviceQueryBase CreateDeviceQuery();
         protected abstract AndroidLogcatSettings LoadEditorSettings();
         protected abstract AndroidTools CreateAndroidTools();
@@ -120,9 +121,9 @@ namespace Unity.Android.Logcat
             m_SubscribeToEditorUpdate = subscribeToEditorUpdate;
         }
 
-        public override AndroidLogcatMessageProviderBase CreateMessageProvider(AndroidBridge.ADB adb, string filter, Priority priority, int packageID, string logPrintFormat, IAndroidLogcatDevice device, Action<string> logCallbackAction)
+        public override AndroidLogcatMessageProviderBase CreateMessageProvider(AndroidBridge.ADB adb, Priority priority, int packageID, string logPrintFormat, IAndroidLogcatDevice device, Action<string> logCallbackAction)
         {
-            return new AndroidLogcatMessageProvider(adb, filter, priority, packageID, logPrintFormat, device, logCallbackAction);
+            return new AndroidLogcatMessageProvider(adb, priority, packageID, logPrintFormat, device, logCallbackAction);
         }
 
         public override void Initialize()