+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatStyles.cs
@@ -3,6 +3,10 @@ using UnityEditor;
 
 namespace Unity.Android.Logcat
 {
+    /// <summary>
+    /// UI styles.
+    /// Note: Don't use EditorStyles, since they're not initialized if Editor is ran in batch mode, for ex., tests
+    /// </summary>
     static class AndroidLogcatStyles
     {
         public const int kFontSize = 10;
@@ -24,7 +28,7 @@ namespace Unity.Android.Logcat
         public static GUIStyle infoSmallStyle = new GUIStyle("CN EntryInfoIconSmall") { fixedHeight = kLogEntryFixedHeight };
         public static GUIStyle warningSmallStyle = new GUIStyle("CN EntryWarnIconSmall") { fixedHeight = kLogEntryFixedHeight };
         public static GUIStyle errorSmallStyle = new GUIStyle("CN EntryErrorIconSmall") { fixedHeight = kLogEntryFixedHeight };
-        public static GUIStyle priorityDefaultStyle = new GUIStyle(EditorStyles.miniLabel)
+        public static GUIStyle priorityDefaultStyle = new GUIStyle("miniLabel")
         {
             fontSize = kLogEntryFontSize,
             fixedHeight = kLogEntryFixedHeight,
@@ -53,10 +57,10 @@ namespace Unity.Android.Logcat
         public static GUIStyle tagEntryBackground = new GUIStyle("CN EntryBackodd") { fixedHeight = kTagEntryFixedHeight };
         public static GUIStyle tagEntryBackgroundOdd = new GUIStyle("CN EntryBackodd") { fixedHeight = kTagEntryFixedHeight };
         public static GUIStyle tagEntryBackgroundEven = new GUIStyle("CN EntryBackEven") { fixedHeight = kTagEntryFixedHeight };
-        public static GUIStyle tagEntryStyle = new GUIStyle(EditorStyles.miniLabel) { fontSize = kTagEntryFontSize, fixedHeight = kTagEntryFixedHeight };
-        public static GUIStyle tagToggleStyle = new GUIStyle(EditorStyles.toggle) { fixedWidth = ktagToggleFixedWidth, fixedHeight = kTagEntryFixedHeight };
-        public static GUIStyle tagButtonStyle = new GUIStyle(EditorStyles.miniButton) { fixedHeight = kTagEntryFixedHeight };
-        public static GUIStyle removeTextStyle = new GUIStyle(EditorStyles.miniLabel) { fontSize = 8, fixedHeight = kTagEntryFixedHeight };
+        public static GUIStyle tagEntryStyle = new GUIStyle("miniLabel") { fontSize = kTagEntryFontSize, fixedHeight = kTagEntryFixedHeight };
+        public static GUIStyle tagToggleStyle = new GUIStyle("toggle") { fixedWidth = ktagToggleFixedWidth, fixedHeight = kTagEntryFixedHeight };
+        public static GUIStyle tagButtonStyle = new GUIStyle("miniButton") { fixedHeight = kTagEntryFixedHeight };
+        public static GUIStyle removeTextStyle = new GUIStyle("miniLabel") { fontSize = 8, fixedHeight = kTagEntryFixedHeight };
         public static readonly GUIStyle kSeriesLabel = "ProfilerPaneSubLabel";
 
         public static GUIStyle stacktraceStyle = new GUIStyle("textArea")
@@ -98,7 +102,7 @@ namespace Unity.Android.Logcat
 
         public static Font GetFont()
         {
-            return (Font)EditorGUIUtility.LoadRequired("Packages/com.unity.mobile.android-logcat/Editor/Resources/consola.ttf");
+            return (Font)EditorGUIUtility.LoadRequired("Packages/com.unity.mobile.android-logcat/Editor/Fonts/consola.ttf");
         }
 
         internal class StatusWheel
@@ -137,7 +141,7 @@ namespace Unity.Android.Logcat
 
         public static Texture2D LoadIcon(string name)
         {
-            var tex = (Texture2D)EditorGUIUtility.Load($"Packages/com.unity.mobile.android-logcat/Editor/Resources/{name}.png");
+            var tex = (Texture2D)EditorGUIUtility.Load($"Packages/com.unity.mobile.android-logcat/Editor/Textures/{name}.png");
             return tex;
         }