+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatUserSettings.cs
@@ -10,8 +10,6 @@ namespace Unity.Android.Logcat
     [Serializable]
     internal class AndroidLogcatUserSettings
     {
-        internal const int kMaxExitedPackages = 5;
-
         [Serializable]
         internal class ScreenRecorderSettings
         {
@@ -45,9 +43,7 @@ namespace Unity.Android.Logcat
         [SerializeField]
         private AndroidLogcatMemoryViewerState m_MemoryViewerState;
         [SerializeField]
-        private string m_Filter;
-        [SerializeField]
-        private bool m_FilterIsRegularExpression;
+        private FilterOptions m_FilterOptions;
         [SerializeField]
         private List<ReordableListItem> m_SymbolPaths;
         [SerializeField]
@@ -138,7 +134,7 @@ namespace Unity.Android.Logcat
             return packages;
         }
 
-        public void CleanupDeadPackagesForDevice(IAndroidLogcatDevice device)
+        public void CleanupDeadPackagesForDevice(IAndroidLogcatDevice device, int maxExitedPackagesToShow)
         {
             if (device == null)
                 return;
@@ -156,7 +152,7 @@ namespace Unity.Android.Logcat
             }
 
             // Need to remove the package which were added first, since they are the oldest packages
-            int deadPackagesToRemove = deadPackageCount - kMaxExitedPackages;
+            int deadPackagesToRemove = deadPackageCount - maxExitedPackagesToShow;
             if (deadPackagesToRemove <= 0)
                 return;
 
@@ -179,6 +175,12 @@ namespace Unity.Android.Logcat
         {
             if (pid <= 0)
                 return null;
+
+            if (device == null)
+            {
+                Debug.LogError("Cannot create package information, since there's no Android device connected.");
+                return null;
+            }
 
             var packages = GetOrCreatePackagesForDevice(device);
             PackageInformation info = packages.FirstOrDefault(package => package.processId == pid);
@@ -238,30 +240,17 @@ namespace Unity.Android.Logcat
             }
         }
 
-        public string Filter
+        public FilterOptions FilterOptions
         {
             set
             {
-                m_Filter = value;
+                m_FilterOptions = value;
             }
             get
             {
-                return m_Filter;
+                return m_FilterOptions;
             }
         }
-
-        public bool FilterIsRegularExpression
-        {
-            set
-            {
-                m_FilterIsRegularExpression = value;
-            }
-            get
-            {
-                return m_FilterIsRegularExpression;
-            }
-        }
-
         public List<ReordableListItem> SymbolPaths
         {
             get => m_SymbolPaths;
@@ -280,6 +269,7 @@ namespace Unity.Android.Logcat
             m_KnownPackages = new Dictionary<string, List<PackageInformation>>();
             m_MemoryViewerState = new AndroidLogcatMemoryViewerState();
             m_SymbolPaths = new List<ReordableListItem>();
+            m_FilterOptions = new FilterOptions();
 
             m_ScreenRecorderSettings = new ScreenRecorderSettings
             {
@@ -291,8 +281,7 @@ namespace Unity.Android.Logcat
                 VideoSizeX = 1280,
                 VideoSizeY = 720,
                 DisplayId = string.Empty
-            };
-        }
+            }        }
 
         internal static AndroidLogcatUserSettings Load(string path)
         {